Refactor all DiffString datasets and models into DiffToken
