+++ common/evaluation.py
@@ -1,4 +1,4 @@
-from common.evaluators.diff_string_evaluator import DiffStringEvaluator
+from common.evaluators.diff_token_evaluator import DiffTokenEvaluator
 from common.evaluators.paired_token_evaluator import PairedTokenEvaluator
 
 
@@ -7,7 +7,7 @@ class EvaluatorFactory(object):
     Get the corresponding Evaluator class for a particular dataset.
     """
     evaluator_map = {
-        'VulasDiff': DiffStringEvaluator,
+        'VulasDiffToken': DiffTokenEvaluator,
         'VulasPairedToken': PairedTokenEvaluator
 
     }
