+++ models/diff_token/reg_cnn/README.md
