+++ models/diff_token/reg_lstm/model.py
@@ -4,8 +4,8 @@ import torch.nn as nn
 
 import torch.nn.functional as F
 
-from models.diff_string.reg_lstm.weight_drop import WeightDrop
-from models.diff_string.reg_lstm.embed_regularize import embedded_dropout
+from models.diff_token.reg_lstm.weight_drop import WeightDrop
+from models.diff_token.reg_lstm.embed_regularize import embedded_dropout
 
 
 class RegLSTM(nn.Module):
